Move get_version to algo/util.py; fix version to 0.25
