+++ pr_agent/cli.py
@@ -3,9 +3,9 @@ import asyncio
 import os
 
 from pr_agent.agent.pr_agent import PRAgent, commands
+from pr_agent.algo.utils import get_version
 from pr_agent.config_loader import get_settings
 from pr_agent.log import get_logger, setup_logger
-from pr_agent.version import get_version
 
 log_level = os.environ.get("LOG_LEVEL", "INFO")
 setup_logger(log_level)
