+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pr-agent"
-version = "0.2.6"
+version = "0.2.5"
 
 authors = [{ name = "CodiumAI", email = "tal.r@codium.ai" }]
